mi_chatbot_app_online/requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
gunicorn==20.1.0
azure-ai-language-questionanswering==X.Y.Z